Clean up the test code to use multiple files
