--- dev_deps.ts
@@ -1,11 +1,0 @@
-export {
-  bench,
-  runBenchmarks,
-} from "https://deno.land/std@0.135.0/testing/bench.ts";
-export {
-  assertAlmostEquals,
-  assertEquals,
-  assertInstanceOf,
-  assertMatch,
-  assertThrows,
-} from "https://deno.land/std@0.135.0/testing/asserts.ts";
